Refactor example app structure and implement navigation; create HomePage, SingleInstancePage, and MultipleInstancesPage. Consolidate locale handling and improve NodeEditorInstance functionality. Remove unused imports and streamline clipboard and rendering utilities.
