+++ lib/src/constants.dart
@@ -1,6 +1,3 @@
-import 'package:flutter/material.dart';
-
-final kNodeEditorWidgetKey = GlobalKey();
 const kMaxEventUndoHistory = 128;
 const kMaxEventRedoHistory = 128;
 const kSpatialHashingCellSize = 2056.0;
