+++ lib/src/core/controller/core.dart
@@ -1,6 +1,5 @@
 import 'dart:math';
 
-import 'package:fl_nodes/src/constants.dart';
 import 'package:fl_nodes/src/core/controller/callback.dart';
 import 'package:fl_nodes/src/core/controller/history.dart';
 import 'package:fl_nodes/src/core/controller/project.dart';
@@ -30,6 +29,7 @@ import 'utils.dart';
 /// sending and receiving events.
 class FlNodeEditorController with ChangeNotifier {
   FlCallback? onCallback;
+  GlobalKey editorKey;
 
   FlNodeEditorController({
     this.config = const FlNodeEditorConfig(),
@@ -38,7 +38,8 @@ class FlNodeEditorController with ChangeNotifier {
     ProjectLoader? projectLoader,
     ProjectCreator? projectCreator,
     this.onCallback,
-  }) {
+    GlobalKey? editorKey,
+  }) : editorKey = editorKey ?? GlobalKey() {
     clipboard = FlNodeEditorClipboard(this);
     runner = FlNodeEditorRunner(this);
     history = FlNodeEditorHistory(this);
@@ -1047,8 +1048,7 @@ class FlNodeEditorController with ChangeNotifier {
       margin: 256,
     );
 
-    final nodeEditorSize =
-        RenderBoxUtils.getSizeFromGlobalKey(kNodeEditorWidgetKey)!;
+    final nodeEditorSize = RenderBoxUtils.getSizeFromGlobalKey(editorKey)!;
 
     setViewportOffset(
       -encompassingRect.center,
