+++ lib/src/widgets/default_node.dart
@@ -67,6 +67,7 @@ class _DefaultNodeWidgetState extends State<DefaultNodeWidget> {
 
   double get viewportZoom => widget.controller.viewportZoom;
   Offset get viewportOffset => widget.controller.viewportOffset;
+  GlobalKey get editorKey => widget.controller.editorKey;
 
   @override
   void initState() {
@@ -148,7 +149,7 @@ class _DefaultNodeWidgetState extends State<DefaultNodeWidget> {
     const edgeThreshold = 50.0;
     final moveAmount = 5.0 / widget.controller.viewportZoom;
     final editorBounds =
-        RenderBoxUtils.getEditorBoundsInScreen(kNodeEditorWidgetKey);
+        RenderBoxUtils.getEditorBoundsInScreen(widget.controller.editorKey);
     if (editorBounds == null) return;
 
     _edgeTimer?.cancel();
@@ -184,8 +185,12 @@ class _DefaultNodeWidgetState extends State<DefaultNodeWidget> {
   }
 
   _TempLink? _isNearPort(Offset position) {
-    final worldPosition =
-        RenderBoxUtils.screenToWorld(position, viewportOffset, viewportZoom);
+    final worldPosition = RenderBoxUtils.screenToWorld(
+      editorKey,
+      position,
+      viewportOffset,
+      viewportZoom,
+    );
 
     final near = Rect.fromCenter(
       center: worldPosition!,
@@ -214,8 +219,12 @@ class _DefaultNodeWidgetState extends State<DefaultNodeWidget> {
   }
 
   void _onTmpLinkUpdate(Offset position) {
-    final worldPosition =
-        RenderBoxUtils.screenToWorld(position, viewportOffset, viewportZoom);
+    final worldPosition = RenderBoxUtils.screenToWorld(
+      editorKey,
+      position,
+      viewportOffset,
+      viewportZoom,
+    );
     final node = widget.controller.nodes[_tempLink!.nodeId]!;
     final port = node.ports[_tempLink!.portId]!;
     final absolutePortOffset = node.offset + port.offset;
@@ -506,8 +515,12 @@ class _DefaultNodeWidgetState extends State<DefaultNodeWidget> {
       );
     }
 
-    final worldPosition =
-        RenderBoxUtils.screenToWorld(position, viewportOffset, viewportZoom);
+    final worldPosition = RenderBoxUtils.screenToWorld(
+      editorKey,
+      position,
+      viewportOffset,
+      viewportZoom,
+    );
 
     return compatiblePrototypes.map((entry) {
       return MenuItem(
